Remove tests for before calling canEvaluatePolicy(_:error:)
Apple says iOS 11.0 would crash if you call biometryType.
It means we can't use it directly.
And LAContext doesn't have any information about whether
the method was called or not.
So we can't return correct biometryType before calling
canEvaluatePolicy(_:error:).

But biometryType doesn't make sense in that case in the first place.
So I removed this testcase.